change for external usage

+++ day_0/ex00/SD.go
@@ -2,11 +2,11 @@ package main
 
 import "math"
 
-func Deviation(inputArray []int) float64 {
+func Deviation(inputArray []float64) float64 {
 	var result float64
 	mean := Mean(inputArray)
 	for _, num := range inputArray {
-		result = result + math.Pow(float64(num)-mean, 2)
+		result = result + math.Pow(num-mean, 2)
 	}
 	return math.Sqrt(result / float64(len(inputArray)))
 }
