+++ day_0/ex00/mean.go
@@ -1,9 +1,9 @@
 package main
 
-func Mean(inputArray []int) float64 {
+func Mean(inputArray []float64) float64 {
 	var result float64
 	for _, num := range inputArray {
-		result = result + float64(num)
+		result = result + num
 	}
 	return result / float64(len(inputArray))
 }
